# libs/shared/apis/v1/synaptiq-apis/src/main/java/com/synaptiq/infrastructure/in/web/api/ActionsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.synaptiq.infrastructure.in.web.api;

import com.synaptiq.infrastructure.in.web.dto.ActionResultResponse;
import com.synaptiq.infrastructure.in.web.dto.ExecuteActionRequest;
import org.springframework.lang.Nullable;
import com.synaptiq.infrastructure.in.web.dto.SavedItemListResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-05T21:15:25.464614100-05:00[America/Chicago]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "Actions", description = "Actions — execution, saved items")
public interface ActionsApi {

    String PATH_EXECUTE_ACTION = "/api/v1/actions/execute";
    /**
     * POST /api/v1/actions/execute : Execute an action
     *
     * @param executeActionRequest  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Action result (status code 200)
     */
    @Operation(
        operationId = "executeAction",
        summary = "Execute an action",
        tags = { "Actions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Action result", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ActionResultResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ActionsApi.PATH_EXECUTE_ACTION,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    Mono<ResponseEntity<ActionResultResponse>> executeAction(
        @Parameter(name = "ExecuteActionRequest", description = "", required = true) @Valid @RequestBody Mono<ExecuteActionRequest> executeActionRequest,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_LIST_SAVED_ITEMS = "/api/v1/actions/saved-items";
    /**
     * GET /api/v1/actions/saved-items : List saved items
     *
     * @param sessionId  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Saved items list (status code 200)
     */
    @Operation(
        operationId = "listSavedItems",
        summary = "List saved items",
        tags = { "Actions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Saved items list", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SavedItemListResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ActionsApi.PATH_LIST_SAVED_ITEMS,
        produces = { "application/json" }
    )
    Mono<ResponseEntity<SavedItemListResponse>> listSavedItems(
        @NotNull @Parameter(name = "sessionId", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "sessionId", required = true) String sessionId,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_REMOVE_SAVED_ITEM = "/api/v1/actions/saved-items/{itemId}";
    /**
     * DELETE /api/v1/actions/saved-items/{itemId} : Remove a saved item
     *
     * @param itemId  (required)
     * @param sessionId  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Item removed (status code 204)
     */
    @Operation(
        operationId = "removeSavedItem",
        summary = "Remove a saved item",
        tags = { "Actions" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Item removed")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ActionsApi.PATH_REMOVE_SAVED_ITEM
    )
    Mono<ResponseEntity<Void>> removeSavedItem(
        @Parameter(name = "itemId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("itemId") String itemId,
        @NotNull @Parameter(name = "sessionId", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "sessionId", required = true) String sessionId,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );

}