# libs/shared/openapi-spec/src/main/resources/openapi/bundled-synaptiq-v1.yaml
openapi: 3.1.0
info:
  title: Synaptiq API
  description: |
    AI-Powered Conversational Commerce Platform — REST API

    Synaptiq enables businesses to deploy intelligent, context-aware AI shopping assistants
    with full catalog management, conversation orchestration, workflow automation,
    and multi-tenant governance. This specification covers all bounded contexts:
    Authentication, Tenants, Catalog, Chat, Actions, Workflows, Analytics,
    Branding, Tenant Configuration, and Schema Registry.
  version: 1.0.0
  contact:
    name: Synaptiq Team
    url: https://synaptiq.dev
  license:
    name: Proprietary
    url: https://synaptiq.dev/license
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.synaptiq.dev
    description: Production server
...
paths:
...
  /api/v1/actions/execute:
    post:
      operationId: executeAction
      summary: Execute an action
      tags:
        - Actions
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteActionRequest'
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResultResponse'
  /api/v1/actions/saved-items:
    get:
      operationId: listSavedItems
      summary: List saved items
      tags:
        - Actions
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Saved items list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedItemListResponse'
  /api/v1/actions/saved-items/{itemId}:
    delete:
      operationId: removeSavedItem
      summary: Remove a saved item
      tags:
        - Actions
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: itemId
          in: path
          required: true
          schema:
            type: string
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Item removed
...
  /api/v1/config/ai:
    get:
      operationId: getAiConfig
      summary: Get ai config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Config data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiPersonaResponse'
    patch:
      operationId: updateAiConfig
      summary: Update ai config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiPersonaResponse'
      responses:
        '200':
          description: Config updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiPersonaResponse'
  /api/v1/config/guardrails:
    get:
      operationId: getGuardrailsConfig
      summary: Get guardrails config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Config data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuardrailsResponse'
    patch:
      operationId: updateGuardrailsConfig
      summary: Update guardrails config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardrailsResponse'
      responses:
        '200':
          description: Config updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuardrailsResponse'
  /api/v1/config/llm:
    get:
      operationId: getLlmConfig
      summary: Get llm config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Config data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmProviderResponse'
    patch:
      operationId: updateLlmConfig
      summary: Update llm config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmProviderResponse'
      responses:
        '200':
          description: Config updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmProviderResponse'
  /api/v1/config/components:
    get:
      operationId: getComponentsConfig
      summary: Get components config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Config data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentsToggleResponse'
    patch:
      operationId: updateComponentsConfig
      summary: Update components config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentsToggleResponse'
      responses:
        '200':
          description: Config updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentsToggleResponse'
  /api/v1/config/actions:
    get:
      operationId: getActionsConfig
      summary: Get actions config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Config data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionsConfigResponse'
    patch:
      operationId: updateActionsConfig
      summary: Update actions config
      tags:
        - Config
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionsConfigResponse'
      responses:
        '200':
          description: Config updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionsConfigResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: The tenant identifier for multi-tenant isolation
      schema:
        type: string
        minLength: 1
        maxLength: 100
  schemas:
...
    ExecuteActionRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
        itemId:
          type: string
        sessionId:
          type: string
        message:
          type: string
    ActionResultResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties: true
    SavedItemResponse:
      type: object
      properties:
        id:
          type: string
        itemId:
          type: string
        sessionId:
          type: string
        itemSnapshot:
          $ref: '#/components/schemas/ItemSnapshotResponse'
        createdAt:
          type: string
          format: date-time
    SavedItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SavedItemResponse'
        total:
          type: integer
...
    AiPersonaResponse:
      type: object
      properties:
        displayName:
          type: string
        tone:
          type: string
        customInstruction:
          type: string
        welcomeMessage:
          type: string
        starterPrompts:
          type: array
          items:
            type: string
    GuardrailsResponse:
      type: object
      properties:
        outOfScopeMessage:
          type: string
        recommendationMode:
          type: boolean
        language:
          type: string
    LlmProviderResponse:
      type: object
      properties:
        provider:
          type: string
        modelId:
          type: string
        isByok:
          type: boolean
    ComponentsToggleResponse:
      type: object
      properties:
        itemCard:
          type: boolean
        itemGrid:
          type: boolean
        itemDetail:
          type: boolean
        comparisonTable:
          type: boolean
        filterSummary:
          type: boolean
        resultCount:
          type: boolean
        emptyState:
          type: boolean
        actionConfirm:
          type: boolean
        infoBanner:
          type: boolean
    ActionsConfigResponse:
      type: object
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ActionEntryResponse'
        enquiryWebhookUrl:
          type: string
        enquiryEmail:
          type: string
    ActionEntryResponse:
      type: object
      properties:
        actionId:
          type: string
        enabled:
          type: boolean
        label:
          type: string
...
    ItemSnapshotResponse:
      type: object
      description: Snapshot of a saved item at the time it was bookmarked
      properties:
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        description:
          type: string
        price:
          type: string
        category:
          type: string
        dataSourceId:
          type: string
        sourceItemId:
          type: string
...